# Use FastSS for fast kNN over Levenshtein distance