… gensim/similarities/docsim.py
… gensim/similarities/termsim.py
… gensim/test/test_similarities.py
… gensim/utils.py