# Introduction

The `LevenshteinSimilarityIndex` term similarity index in the `termsim.levenshtein` module implements the lexical text similarity search technique described by [Charlet and Damnati (2017)](https://www.aclweb.org/anthology/S17-2051/) in their paper describing their winning system at SemEval-2017 Task 3: Community Question Answering.

We are showing a related semantic similarity search technique using the `WordEmbeddingSimilarityIndex` term similarity index in our [Soft Cosine Similarity autoexample][1], which enjoys some popularity among our users. We would like to also advertise `LevenshteinSimilarityIndex`, which provides a different but equally useful kind of search. However, the current implementation uses brute-force kNN search over the Levenshtein distance to produce a term similarity matrix, which is so slow that it can take years to produce a matrix even for medium-sized corpora such as the English Wikipedia.

Following the discussion in https://github.com/RaRe-Technologies/gensim/issues/2541, @piskvorky and I implemented indexing using [the FastSS algorithm][2] for kNN search over the Levenshtein distance in hopes that this would speed `LevenshteinSimilarityIndex` up by at least three orders of magnitude (1,000×), so that it can compete with `WordEmbeddingSimilarityIndex`. As an added bonus, using the FastSS algorithm allows us to remove our external dependence on the Levenshtein library.

 [1]: https://radimrehurek.com/gensim/auto_examples/tutorials/run_scm.html
 [2]: https://fastss.csg.uzh.ch/ifi-2007.02.pdf

# Speed comparison
Below, I will show a before-and-after speed comparison of `LevenshteinSimilarityIndex` compared to the standard `WordEmbeddingSimilarityIndex` shown in [the Soft Cosine Similarity autoexample][1]. We are measuring how many kNN searches per second, k = 100, a term similarity index can perform. To produce my dictionary (253,854 terms) and word embeddings, I will use [the text8 corpus][4] (100 MB). I am running the code on a [Dell Inspiron 15 7559][3].

 [3]: https://www.notebookcheck.net/Dell-Inspiron-15-7559-Notebook-Review.154635.0.html
 [4]: http://mattmahoney.net/dc/textdata

## Before the change

We can see that even with our tiny corpus, the `LevenshteinSimilarityIndex` takes **over three days** to find the 100 nearest neighbors for all 253,854 terms in our vocabulary. Contrast this with the `WordEmbeddingSimilarityIndex`, which finishes in under four minutes even though we are using exact nearest-neighbor search and we could get further speed-up by using e.g. [the Annoy index][5].

 [5]: https://radimrehurek.com/gensim/similarities/termsim.html

``` python
$ pip install gensim==4.0.1 python-Levenshtein
$ wget http://mattmahoney.net/dc/text8.zip
$ unzip text8.zip
$ python
>>> from gensim.corpora import Dictionary
>>> from gensim.models.word2vec import LineSentence, Word2Vec
>>> from gensim.similarities import (
...     SparseTermSimilarityMatrix,
...     WordEmbeddingSimilarityIndex,
...     LevenshteinSimilarityIndex,
... )
>>> 
>>> corpus = LineSentence('text8')
>>> dictionary = Dictionary(corpus)
>>> w2v_model = Word2Vec(sentences=corpus)
>>> embedding_index = WordEmbeddingSimilarityIndex(w2v_model.wv)
>>> levenshtein_index = LevenshteinSimilarityIndex(dictionary)
>>>
>>> SparseTermSimilarityMatrix(embedding_index, dictionary)
100%|███████████████████████████████| 253854/253854 [04:04<00:00, 1037.97it/s]
>>> SparseTermSimilarityMatrix(levenshtein_index, dictionary)
  0%|                               | 124/253854 [02:24<80:18:05,  1.14s/it]
```

## After the change

With the FastSS algorithm, the `LevenshteinSimilarityIndex` receives a 1,500× speed-up and is now not only not slower than the `WordEmbeddingSimilarityIndex`, but 1.5× faster. Both term similarity indexes now find the 100 nearest neighbors for all 253,854 terms in our vocabulary in under an hour.
``` python
$ pip install lexpy git+https://github.com/witiko/gensim@7054f90
$ python
>>> from gensim.corpora import Dictionary
>>> from gensim.models.word2vec import LineSentence, Word2Vec
>>> from gensim.similarities import (
...     SparseTermSimilarityMatrix,
...     WordEmbeddingSimilarityIndex,
...     LevenshteinSimilarityIndex,
... )
>>> 
>>> corpus = LineSentence('text8')
>>> dictionary = Dictionary(corpus)
>>> w2v_model = Word2Vec(sentences=corpus)
>>> embedding_index = WordEmbeddingSimilarityIndex(w2v_model.wv)
>>> levenshtein_index = LevenshteinSimilarityIndex(dictionary)
>>>
>>> SparseTermSimilarityMatrix(embedding_index, dictionary)
100%|███████████████████████████████| 253854/253854 [03:57<00:00, 1070.14it/s]
>>> SparseTermSimilarityMatrix(levenshtein_index, dictionary)
100%|███████████████████████████████| 253854/253854 [02:34<00:00, 1639.23it/s]
```

# Conclusion
Using the FastSS algorithm for kNN search over the Levenshtein distance, we managed to increase the speed of the `LevenshteinSimilarityIndex` term similarity index by four orders of magnitude (1,500×) on the text8 corpus. As an added bonus, using the FastSS algorithm allowed us to remove our external dependence on the Levenshtein library. Closes #2541.