@piskvorky 2021-05-15
@Witiko what's the result of the benchmark, what am I looking at here?

```python
>>> SparseTermSimilarityMatrix(embedding_index, dictionary)
100%|████████████████████████████| 10781/10781 [00:01<00:00, 6401.08it/s]
```