@Witiko 2021-05-15
We are going over the whole dictionary and looking for the 100 nearest neighbors for every word. Your output shows that `WordEmbeddingSimilarityIndex` is able to do 6400 kNN searches per second.