@Witiko 2021-05-16
Following our discussion in https://github.com/RaRe-Technologies/gensim/issues/2541#issuecomment-841724918, I implemented the DAWG structure for approximate kNN search over the Levenshtein distance.

# Speed comparison
Let's see how the different techniques measure up. This time, I used the ca 200× larger [text8 dataset (100 MB)][1], which should be more representative. The rest of the experimental setup is unaltered from https://github.com/RaRe-Technologies/gensim/pull/3146#issue-645217217.

 [1]: http://mattmahoney.net/dc/textdata

## Brute-force kNN search

``` python
$ pip install gensim==4.0.1 python-Levenshtein
$ wget http://mattmahoney.net/dc/text8.zip
$ unzip text8.zip
$ python
>>> from gensim.corpora import Dictionary
>>> from gensim.models.word2vec import LineSentence, Word2Vec
>>> from gensim.similarities import (
...     SparseTermSimilarityMatrix,
...     WordEmbeddingSimilarityIndex,
...     LevenshteinSimilarityIndex,
... )
>>> 
>>> corpus = LineSentence('text8')
>>> dictionary = Dictionary(corpus)
>>> w2v_model = Word2Vec(sentences=corpus)
>>> embedding_index = WordEmbeddingSimilarityIndex(w2v_model.wv)
>>> levenshtein_index = LevenshteinSimilarityIndex(dictionary)
>>>
>>> SparseTermSimilarityMatrix(embedding_index, dictionary)
  4%|████▌                       | 9949/253854 [00:33<13:51, 293.16it/s]
>>> SparseTermSimilarityMatrix(levenshtein_index, dictionary)
  0%|                            | 20/253854 [00:22<79:30:25,  1.13s/it]
```

## VP-Tree index

``` python
$ pip install vptree git+https://github.com/witiko/gensim@af5833d
$ python
>>> # Same as above
>>>
>>> SparseTermSimilarityMatrix(embedding_index, dictionary)
  4%|████▌                       | 9949/253854 [00:33<13:51, 293.16it/s]
>>> SparseTermSimilarityMatrix(levenshtein_index, dictionary)
  0%|                            | 20/253854 [00:17<61:20:21,  1.15it/s]
```

## DAWG index

``` python
$ pip install lexpy git+https://github.com/witiko/gensim@e24e3e90
$ python
>>> # Same as above
>>>
>>> SparseTermSimilarityMatrix(embedding_index, dictionary)
  4%|████▌                       | 9949/253854 [00:33<13:51, 293.16it/s]
>>> SparseTermSimilarityMatrix(levenshtein_index, dictionary)
  4%|████▌                       | 9993/253854 [01:55<44:54, 90.50it/s]
```

# Conclusion

Compared to the brute force kNN and the VP-Tree metric index, DAWG approaches the speed of the word embedding kNN retrieval at the expense of only finding the terms with the Levenshtein distance 1 from the query.