@piskvorky 2021-05-16
@Witiko I tried with automata (the first Python implementation that [Google gave me](https://github.com/antoinewdg/pyffs)):

```
max_distance=1:
100%|████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████| 10781/10781 [00:34<00:00, 310.78it/s]
2021-05-16 03:38:35,834 : INFO : constructed a sparse term similarity matrix with 0.025356% density
CPU times: user 34.6 s, sys: 65.1 ms, total: 34.6 s
Wall time: 34.7 s

max_distance=2:
█████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████| 10781/10781 [01:57<00:00, 91.82it/s]
2021-05-16 03:35:15,161 : INFO : constructed a sparse term similarity matrix with 0.124872% density
CPU times: user 1min 56s, sys: 283 ms, total: 1min 56s
Wall time: 1min 57s
```

The automaton implementation seems pure-Python, so I expect 10-100x speed-up after compiling / optimizing.

Am I missing something? Why do you consider this problem so hard?