@Witiko 2021-05-16
> Which doesn't seem too bad… am I missing something?

Very nice! The DAWG implementation from lexpy is also pure Python, but your implementation is a fait bit faster! 🙂 

> Why do you consider this problem so hard?

In my experience, kNN over the Levenshtein distance is difficult to optimize unless you are willing to cut corners, which we are doing right now by limiting the number of results to a small radius.

> I expect 10-100x speed-up after compiling / optimizing.

Sadly, I can't volunteer the time. The best I can offer is code with an external dependency.