@Witiko 2021-05-16
> [...] your implementation is a fait bit faster! 🙂

@piskvorky Actually, you tried comparing your speed with the `WordEmbeddingSimilarityIndex`?
You may well be using a faster machine and our results may be incomparable.