@piskvorky 2021-05-16
> I regret I can't volunteer the time. The best I can offer is code with an external dependency.

A pair of eyes to sanity-check the solution would be even better :) I just plugged the code in, didn't even look at the results: 5bd8bf85115fbb8066789762caa4b5ed8365ac1c

@Witiko could you double check 5bd8bf85115fbb8066789762caa4b5ed8365ac1c please? You're much more involved in this code than me.

> Sadly, it does not seem published to pypi, so it's difficult to use as a dependency and we would need to adopt it.

That particular automaton lib has an MIT license, so IMO it'd make more sense to include it in Gensim directly, rather than as external dependency. Especially since it's not even a package / on PyPI.

But two points:

1. How important is this functionality? What are its use-cases, motivation, user base? Basically, is it worth all the hassle.
2. I'd hope a more thorough googling will reveal a better (faster, packaged, maintained) implementation. Or failing that, we could code it up ourselves – I didn't check in detail but the algo seems straightforward: prebuild a trie of the static dictionary and then intersect it dynamically with each automation built from the query string.

On the theoretical side, we should be able to do even better, because this algo doesn't take into account the fact that the query words are the words from the dictionary itself. But that seems a very useful fact. This algo is more general, the query can be anything, but that also means the algo is not as optimal as it could be.

But before the theoretical algo optimization, let's clear up its practical value / potential impact.