@piskvorky 2021-05-16
> Actually, have you tried comparing your speed with the WordEmbeddingSimilarityIndex?

Good idea, to anchor / sync our results.

Here are both results on my MBP: 

![Uploading Screen Shot 2021-05-16 at 12.38.40.png…]()