@Witiko 2021-05-16
@piskvorky Oh, you tested on the Lee corpus. Then the results are not so great. With the DAWG implementation from this PR, I received up to 10x speed-up compared to brute-force on the Lee corpus (see https://github.com/RaRe-Technologies/gensim/issues/2541#issuecomment-841729768), whereas your implementation achieves only ca 4x speed-up compared to brute-force. That is only twice as fast as the VP-Tree.